+++ CHANGELOG.md
@@ -91,6 +91,8 @@
 
 - Added `?download` file query parameter option to instruct the browser to always download a file and not show a preview.
 
+- Added support for wrapped API errors (_in case Go 1.20+ is used with multiple wrapped errors, `apis.ApiError` takes precedence_).
+
 
 ## v0.16.10